+++ grails-app/domain/foodpaint/PurchaseSheetDet.groovy
@@ -61,7 +61,7 @@ class PurchaseSheetDet{
     /*
     * 進貨數量
     */
-    int qty
+    long qty
 
     static mapping = {
         importFlag  defaultValue: -1
